sportswear_shop/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
diff-match-patch==20241021
Django==5.1.4
django-environ==0.12.0
django-import-export==4.3.4
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
executing==2.1.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openpyxl==3.1.5
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.7.0
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
